Update README.md
After discussing with Irene, she told me that "we will not see anything on the front end because most of it shows the GUI elements". Therefore, I erased the copied code from the index so that it does not get confusing. The group project code is located in the index.py file.